refactor: split MobileSettings and SubscriptionManager into memoised sections
Extract each visual section into its own React.memo component so a
toggle in one section only repaints that section, not the entire screen.
MobileSettings (619→91 lines) and SubscriptionManager (619→208 lines)
are now thin orchestrators over memoised sub-components. Each section
owns its own hooks and handlers. Accessibility labels added throughout.

New files:
- SettingRow.tsx, settingsIcons.tsx, settingsOptions.ts (shared primitives)
- AccountSection, AppSection, PrivacySection, DownloadsSection,
  SyncSection, PerformanceSection, AccountActionsSection
- subscription/: CurrentPlanCard, BillingToggle, PlanCard,
  FreePlanCard, RestoreFooter, subscriptionMeta
